Move.cs, Hierarchy
Made Move.cs not forget what we are contacting.
Moved the scripts folder to the assets folder.
