+++ Assets/Scenes/SampleScene.unity
@@ -422,7 +422,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 695367569}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 2.15, y: 4.06, z: -0.30755773}
+  m_LocalPosition: {x: 2.15, y: 4.25, z: -0.30755773}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -442,7 +442,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 5
-  ground: {fileID: 0}
+  ground: {fileID: 1639121304}
   jumpSpeed: 10
 --- !u!1 &705507993
 GameObject:
