--- Assets/Scenes/Scripts/Camera.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Camera : MonoBehaviour
-{
-    public GameObject player;
-    public float xOffset, yOffset, zOffset;
-    // Update is called once per frame
-    void Update()
-    {
-        transform.position = player.transform.position + new Vector3(xOffset, yOffset,zOffset);
-        transform.LookAt(player.transform.position);
-    }
-}
