+++ Assets/Scripts/pickup.cs.meta
